+++ src/service/index.ts
@@ -1,1 +1,1 @@
-export * from './service';
+export * from './Service';
